FA sniffer: Improve error handling, extend options for capture
Extend the ways the start time can be specified using capture to
include a relative time option (up to just under 24 hours) and
improve the ways invalid times and dates are reported.

Fix daft option bug in -o option for capture.

Also add checking for attempts to read past the currently writing
time in the archive.  It is still possible for reading to underrun
without any error being reported, as this is difficult to detect and
notify!

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/drivers/pcie_sniffer@50914 e099a375-04f9-0310-9d5f-a741eaff62e1