FA sniffer: Add gap identification support
Gaps are now identified in captured data by sending a list of gap
timestamps and offsets together with the rest of the data stream.
The main timestamp should now be more accurately reported, also, but
there remains one discrepancy: if we capture contiguous data up to a
gap boundary we still get a gap recorded!

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/drivers/pcie_sniffer@51166 e099a375-04f9-0310-9d5f-a741eaff62e1