FA sniffer: Separate FA sniffer driver from its tools
The FA archiver tools are here being moved into a location suitable
for managing a final installation of these tools, and so that the kernel
driver is separated from its associated toolset.

The chosen location, diamond/trunk/python/fa-archiver, is a little odd,
but this location fits best with the existing release infrastructure,
in particular it should now be possible to make the FA archiver tools
available to everybody with the `dls-release.py` command.


git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/drivers/pcie_sniffer@51483 e099a375-04f9-0310-9d5f-a741eaff62e1