FA sniffer: Fix return codes from read()
Correct error and interrupt handling is a little bit tricky if we're not
to lose data, in particular in the presence of SA_RESTART we could be
silently losing data.  When a signal occurs we can only return -EINTR if
there is no data to return, otherwise all the data already returned is lost.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/drivers/pcie_sniffer@62137 e099a375-04f9-0310-9d5f-a741eaff62e1