FA sniffer: Protect file access with a mutex
Although only one open at a time is allowed per device, it's still possible
for concurrent access, so guard the read() method with a mutex on the
fa_sniffer_open structure.

Also change the return code from read() when running out of data to allow
all the available data to be read -- it's not an error to run out of data!

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/drivers/pcie_sniffer@62136 e099a375-04f9-0310-9d5f-a741eaff62e1