FA sniffer: Fix mutex starvation issue
Unfortunately using a mutex to ensure single threading of access to the
sniffer device has its problems.  If the mutex is not dropped while
waiting for data to arrive (this is clearly incorrect behaviour anyway)
it turns out that the ioctls are starved when the reader is busy.

In this commit mutex is dropped in favour of a lock free implementation.
This is a bit more complicated to manage and requires some interrupt
routine changes, but the alternative of a wait loop with a dropped mutex
is also surprisingly complicated.

Advanced driver to version 1.4.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/drivers/pcie_sniffer@63380 e099a375-04f9-0310-9d5f-a741eaff62e1